requirements.txt
flask==2.3.2
flask-socketio==6.3.3
eventlet==0.33.4
python-dotenv==1.0.0
SQLAlchemy==2.0.19
pymysql==1.0.3
python-socketio==5.9.0
aiohttp==3.8.4
playwright==1.38.0
python-socketio[asyncio_client]==5.9.0
psutil==5.9.5
redis==4.7.0
duckdb==0.8.1